+++ R/class_options.R
@@ -21,7 +21,8 @@ options_init <- function(
   workspaces = NULL,
   workspace_on_error = NULL,
   seed = NULL,
-  controller = NULL
+  controller = NULL,
+  trust_object_timestamps = NULL
 ) {
   options_new(
     tidy_eval = tidy_eval,
@@ -46,7 +47,8 @@ options_init <- function(
     workspaces = workspaces,
     workspace_on_error = workspace_on_error,
     seed = seed,
-    controller = controller
+    controller = controller,
+    trust_object_timestamps = trust_object_timestamps
   )
 }
 
@@ -73,7 +75,8 @@ options_new <- function(
   workspaces = NULL,
   workspace_on_error = NULL,
   seed = NULL,
-  controller = NULL
+  controller = NULL,
+  trust_object_timestamps = NULL
 ) {
   options_class$new(
     tidy_eval = tidy_eval,
@@ -98,7 +101,8 @@ options_new <- function(
     workspaces = workspaces,
     workspace_on_error = workspace_on_error,
     seed = seed,
-    controller = controller
+    controller = controller,
+    trust_object_timestamps = trust_object_timestamps
   )
 }
 
@@ -131,6 +135,7 @@ options_class <- R6::R6Class(
     workspace_on_error = NULL,
     seed = NULL,
     controller = NULL,
+    trust_object_timestamps = NULL,
     initialize = function(
       tidy_eval = NULL,
       packages = NULL,
@@ -154,7 +159,8 @@ options_class <- R6::R6Class(
       workspaces = NULL,
       workspace_on_error = NULL,
       seed = NULL,
-      controller = NULL
+      controller = NULL,
+      trust_object_timestamps = NULL
     ) {
       self$tidy_eval <- tidy_eval
       self$packages <- packages
@@ -179,6 +185,7 @@ options_class <- R6::R6Class(
       self$workspace_on_error <- workspace_on_error
       self$seed <- seed
       self$controller <- controller
+      self$trust_object_timestamps <- trust_object_timestamps
     },
     export = function() {
       list(
@@ -202,7 +209,8 @@ options_class <- R6::R6Class(
         debug = self$get_debug(),
         workspaces = self$get_workspaces(),
         workspace_on_error = self$get_workspace_on_error(),
-        seed = self$get_seed()
+        seed = self$get_seed(),
+        trust_object_timestamps = self$get_trust_object_timestamps()
       )
     },
     import = function(list) {
@@ -227,6 +235,7 @@ options_class <- R6::R6Class(
       self$set_workspaces(list$workspaces)
       self$set_workspace_on_error(list$workspace_on_error)
       self$set_seed(list$seed)
+      self$set_trust_object_timestamps(list$trust_object_timestamps)
     },
     reset = function() {
       self$tidy_eval <- NULL
@@ -252,6 +261,7 @@ options_class <- R6::R6Class(
       self$workspace_on_error <- NULL
       self$seed <- NULL
       self$controller <- NULL
+      self$trust_object_timestamps <- NULL
     },
     get_tidy_eval = function() {
       self$tidy_eval %|||% TRUE
@@ -321,6 +331,9 @@ options_class <- R6::R6Class(
     },
     get_controller = function() {
       self$controller
+    },
+    get_trust_object_timestamps = function() {
+      self$trust_object_timestamps %|||% TRUE
     },
     set_tidy_eval = function(tidy_eval) {
       self$validate_tidy_eval(tidy_eval)
@@ -414,6 +427,10 @@ options_class <- R6::R6Class(
       self$validate_controller(controller)
       self$controller <- controller
     },
+    set_trust_object_timestamps = function(trust_object_timestamps) {
+      self$validate_trust_object_timestamps(trust_object_timestamps)
+      self$trust_object_timestamps <- trust_object_timestamps
+    },
     validate_tidy_eval = function(tidy_eval) {
       tar_assert_scalar(tidy_eval)
       tar_assert_lgl(tidy_eval)
@@ -487,13 +504,16 @@ options_class <- R6::R6Class(
     },
     validate_debug = function(debug) {
       tar_assert_chr(debug)
+      tar_assert_none_na(trust_object_timestamps)
     },
     validate_workspaces = function(workspaces) {
       tar_assert_chr(workspaces)
+      tar_assert_none_na(trust_object_timestamps)
     },
     validate_workspace_on_error = function(workspace_on_error) {
       tar_assert_scalar(workspace_on_error)
       tar_assert_lgl(workspace_on_error)
+      tar_assert_none_na(trust_object_timestamps)
     },
     validate_seed = function(seed) {
       tar_assert_scalar(seed)
@@ -504,6 +524,13 @@ options_class <- R6::R6Class(
     validate_controller = function(controller) {
       if (!is.null(controller)) {
         validate_crew_controller(controller)
+      }
+    },
+    validate_trust_object_timestamps = function(trust_object_timestamps) {
+      if (!is.null(trust_object_timestamps)) {
+        tar_assert_lgl(trust_object_timestamps)
+        tar_assert_scalar(trust_object_timestamps)
+        tar_assert_none_na(trust_object_timestamps)
       }
     },
     validate = function() {
@@ -530,6 +557,7 @@ options_class <- R6::R6Class(
       self$validate_workspace_on_error(self$get_workspace_on_error())
       self$validate_seed(self$get_seed())
       self$validate_controller(self$get_controller())
+      self$validate_trust_object_timestamps(self$get_trust_object_timestamps())
     }
   )
 )